Cleanup show/hide
It's important here that we changed the button selector, because before
you were binding the first button on the entire page.  It was just luck
that it happened to be the one you were looking for.

We now put `display: none` on the box when it's not being animated or
shown.  That way, it won't block things below it from being clicked
(when something is opacity 0 it can't be seen, but it can still be
clicked on).