Don't bother with IE8
We now wait until the DOM is ready to initialize.  That way, we only
have to get a ref to the suggest_box once, instead of in every scroll
event.  Also, we now use a more sane way of binding the scroll event
which won't break if anyone else on the page tries to do it.